Add multiple books as one JSON batch request
For me it was not obvious how I could POST multiple books as part of one request.
I got hinted to the OData specification where I was able to find some information on this.
I would suggest we add a sample $batch request to our prominent bookshop sample, so that
it is easier for beginners to understand this concept.